SAMPLE/face-recognition-app/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
deepface==0.0.79
pymilvus==2.3.4
pillow==10.0.1
numpy>=1.25.2,<2.6.0
scikit-learn==1.3.2
python-dateutil==2.8.2
tensorflow>=2.15.0
